LayoutTests/crypto/webkitSubtle/sha-384-expected.txt

CONSOLE MESSAGE: line 1: WebKitSubtleCrypto is deprecated. Please use SubtleCrypto instead.
Test crypto.webkitSubtle.digest.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


SHA384 of []
PASS bytesToHexString(new Uint8Array(digest)) is '38b060a751ac96384cd9327eb1b1e36a21fdb71114be07434c0cc7bf63f6e1da274edebfe76f65fbd51ad2f14898b95b'
SHA384 of [0x0]
PASS bytesToHexString(new Uint8Array(digest)) is 'bec021b4f368e3069134e012c2b4307083d3a9bdd206e24e5f0d86e13d6636655933ec2b413465966817a9c208a11717'
SHA384 of [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10]
PASS bytesToHexString(new Uint8Array(digest)) is '89bfcf569ae4af718510da78c67414109f5739bb5c40d51c9c8c50e2b2cee86f2f80c8b9d68f7c01201a0714572fe602'
PASS successfullyParsed is true

TEST COMPLETE
